+++ src/pyrepo/inspecting.py
@@ -49,6 +49,7 @@ def inspect_project(dirpath: str | Path | None = None) -> dict:
         # read_configuration().
         # "version": cfg["metadata"].get("version"),
         "keywords": cfg["metadata"].get("keywords", []),
+        "classifiers": cfg["metadata"].get("classifiers", []),
         "supports_pypy3": False,
         "default_branch": git.Git(dirpath=directory).get_default_branch(),
     }
@@ -73,7 +74,7 @@ def inspect_project(dirpath: str | Path | None = None) -> dict:
         env["uses_versioningit"] = False
 
     env["python_versions"] = []
-    for clsfr in cfg["metadata"]["classifiers"]:
+    for clsfr in env["classifiers"]:
         if m := re.fullmatch(r"Programming Language :: Python :: (\d+\.\d+)", clsfr):
             env["python_versions"].append(util.PyVersion.parse(m[1]))
         if clsfr == "Programming Language :: Python :: Implementation :: PyPy":